fix(ocpp): negotiate the ocpp1.6 subprotocol, split oversized modules - Related to #920
_handle_charger() built web.WebSocketResponse(heartbeat=...) with no
protocols=, so the server never selected the ocpp1.6 subprotocol the
charger offers. OCPP-J 1.6 3.1.2 requires the central system to select
it; a client that gets none back may close the connection outright. This
go-e firmware tolerates it, which is why it never presented as a symptom
- but home-assistant.log carried

  Client protocols ['ocpp1.6'] don't overlap server-known ones ()

on almost every connect, from the very first log in this investigation.
It was visible the whole time and went unread. Four fixes were built on
top of a handshake that was quietly non-conformant.

Fixed with protocols=("ocpp1.6",), plus a warning when negotiation
yields anything else - a mismatch there explains a charger that connects
and immediately drops. The regression test drives a real WebSocket and
asserts ws.protocol, because the bug lives entirely in the handshake and
a unit test against the handler cannot see it; verified by reverting the
fix and watching it fail with the exact production log line.

Also fixes two 30 KB limit violations this branch introduced.
ocpp_server.py (31185) and ocpp_commands.py (32025) were both over. Split
along responsibility lines rather than by size:

  ocpp_anti_flap.py  update_charge_target() - continuous target to
                     discrete start/stop commands
  ocpp_profiles.py   charging-profile construction and dispatch, for
                     both the real limit and the 0 A stop

Every module is now well inside the limit. Mixins that define a sender
must precede those carrying a Callable forward declaration of it, or
mypy reports incompatible base-class definitions. ocpp_server.py
re-exports charger_appears_stalled/CHARGER_STALL_THRESHOLD_S via __all__,
since ocpp_sensors.py and coordinator_data.py document them at that path.
